ajout connection

+++ views/produit-article/liste.php
@@ -13,10 +13,12 @@
                 <p class="card-text"><?= $article->contenu ?></p>
                 <p class="card-text">
                     <a class="text-align-right" href="index.php?route=details-article&id=<?= $article->id ?>">Lire l'article</a>
-                    <a class="text-align-right" href="index.php?route=modifier-article&id=<?= $article->id ?>">Modifier</a>
-                    <a class="text-align-right text-danger ml-2" href="index.php?route=supprimer-article&id=<?= $article->id ?>" onclick="return confirm('Etes-vous certains de vouloir supprimer ?')">Supprimer</a>
-            
-
+                    
+                    <?php if (empty($_SESSION['identifiant']) && $_SESSION['pseudo'] == 'admin') : ?>
+                        <a class="text-align-right" href="index.php?route=modifier-article&id=<?= $article->id ?>">Modifier</a>
+                        <a class="text-align-right text-danger ml-2" href="index.php?route=supprimer-article&id=<?= $article->id ?>" onclick="return confirm('Etes-vous certains de vouloir supprimer ?')">Supprimer</a>
+                        <a class="accueil-link" href="index.php?route=ajouter-article">Ajouter un Article</a>
+                    <?php endif; ?>
 
                 </p>
